requirements.txt
Flask==2.3.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30